requirements.txt
django
djangorestframework
djangorestframework-jwt
django-cors-headers
gunicorn==19.9.0